+++ docs/changelog.rst
@@ -9,3 +9,6 @@ Changelog
   look at the
   `Bartacus Standard Edition <https://github .com/Bartacus/Bartacus-Standard>`_
   how to fill your own ``config.yml``.
+* Add the content object as third parameter to user functions from services.
+* Configure Symfony translator with locale from TypoScript setup.
+* Add routing for content elements to controllers.